Add info of current event to displaychannel
Adds the ability to show the information of the current event in the channel display.
If activated, the information will be displayed automatically, or if the Zapcockpit patch
is active in the VDR core, the info can also be displayed with a 2nd OK.
This can be configured in the plugin setup.